addition of actual project in newfolder(2) and all necessary files
I have created a new folder whereby I created a new direction of the project. I added the folder with the relevant html, css, js, geojsonjson files and pngs and csv. The project is now added to a free website to display how dynamic the website is. I had been trialling with lots of errors and not getting anywhere because I did not realise that parts of the geojson/json, js and other dynamic elements would not load unless I used a server. I therefore added node.js http server and created a website to push edited and additional files so I can view changes on the fly. I have been experiment with d3 as well and there are some d3 elements to the website. Although because a lot of the work is based around data visuals on a map it has been easier to use google api maps with geojson and json and other script to display my data. i have used a website that also allows me to add my own and view it from a url which is not a shortcut but a  way of quicken my limited time avialable to me as I am caring for my wife who has a severe hemoragge due to ivf implantation and a successsful preganancy. The website that is showing my data allowed me to therefore present data and create the maps online so that i can add them as urls and therefore limited multiple local files. The next stage is for me to add more interactive data to the homepage map whereby I have successfuly created custom 'markers' that will display more indepth information around the history of migration by economic migration and invasion through regional dna markers known as haplogroups which is passed down either mtDNA or pDNA, maternal or paternal linegages.