Final Project Website with data
This is the final website files and data. With index.html being the main main page. It has been tested and works in the local directory by clicking on index.html, but will not load the europe google api json data this works though in the free website created and also by using node.js. It is obvious now that it needs a http address to load the data on the index.html page. The files have been put w3schools valiator and the only one that shows an issue is the tablejsondata.html where a warning suggests that I use a css style to format the table, although this is in place already. And I have no answer to that. Files have now been placed into relevant folders for example css, script and media folder. To better organise the website files. With the source links in the html and script files adjusted to suit.